Move RefetchFunction and FetchMoreFunction type over to react/internal
