+++ src/react/hooks/useLoadableQuery.ts
@@ -17,7 +17,12 @@ import type {
   SubscribeToMoreFunction,
   SubscribeToMoreOptions,
 } from "@apollo/client/core";
-import type { CacheKey, QueryRef } from "@apollo/client/react/internal";
+import type {
+  CacheKey,
+  FetchMoreFunction,
+  QueryRef,
+  RefetchFunction,
+} from "@apollo/client/react/internal";
 import {
   assertWrappedQueryRef,
   getSuspenseCache,
@@ -33,7 +38,6 @@ import { invariant } from "@apollo/client/utilities/invariant";
 
 import { __use, useRenderGuard } from "./internal/index.js";
 import { useApolloClient } from "./useApolloClient.js";
-import type { FetchMoreFunction, RefetchFunction } from "./useSuspenseQuery.js";
 import { useWatchQueryOptions } from "./useSuspenseQuery.js";
 
 export type LoadQueryFunction<TVariables extends OperationVariables> = (
