+++ src/react/internal/index.ts
@@ -15,4 +15,8 @@ export {
 } from "./cache/QueryReference.js";
 export type { SuspenseCacheOptions } from "./cache/SuspenseCache.js";
 export type { HookWrappers } from "../hooks/internal/wrapHook.js";
-export type { VariablesOption } from "./types.js";
+export type {
+  FetchMoreFunction,
+  RefetchFunction,
+  VariablesOption,
+} from "./types.js";
