+++ src/react/internal/types.ts
@@ -1,4 +1,11 @@
-import type { OperationVariables } from "@apollo/client/core";
+import type {
+  ApolloQueryResult,
+  FetchMoreQueryOptions,
+  MaybeMasked,
+  ObservableQuery,
+  OperationVariables,
+  Unmasked,
+} from "@apollo/client/core";
 import type { OnlyRequiredProperties } from "@apollo/client/utilities";
 
 export type VariablesOption<TVariables extends OperationVariables> =
@@ -16,3 +23,20 @@ export type VariablesOption<TVariables extends OperationVariables> =
       /** {@inheritDoc @apollo/client!QueryOptionsDocumentation#variables:member} */
       variables: TVariables;
     };
+
+export type RefetchFunction<
+  TData,
+  TVariables extends OperationVariables,
+> = ObservableQuery<TData, TVariables>["refetch"];
+
+export type FetchMoreFunction<TData, TVariables extends OperationVariables> = (
+  fetchMoreOptions: FetchMoreQueryOptions<TVariables, TData> & {
+    updateQuery?: (
+      previousQueryResult: Unmasked<TData>,
+      options: {
+        fetchMoreResult: Unmasked<TData>;
+        variables: TVariables;
+      }
+    ) => Unmasked<TData>;
+  }
+) => Promise<ApolloQueryResult<MaybeMasked<TData>>>;
